+++ src/components/calculator/calculatorStyle.scss
